requirements.txt
discord.py
mpmath
sympy
google-api-python-client
google_auth_oauthlib
